Refactor: modified SOLID원칙에 따른 수정, 트랜잭션 수정
- 단일 책임 원칙에 따른 분리
- 불필요한 트랜잭션 제거 및 단일 트랜잭션 적용
- 금액,포인트 변수타입 Long변경

+++ src/main/java/kr/hhplus/be/server/application/order/dto/request/OrderCreateRequest.java
@@ -15,7 +15,7 @@ public record OrderCreateRequest(
     Long couponId,
     
     @Min(value = 0, message = "포인트는 0 이상이어야 합니다")
-    Integer pointToUse,
+    Long pointToUse,
     
     @NotNull(message = "결제 수단은 필수입니다")
     String paymentType,
